Revert "Merge branch 'master' into App_with_BurgerButton"
This reverts commit 6356b1a1215f9e1194a4ed4b946fe233bc75dee6, reversing
changes made to db55a3497dab298fbb654f47f0f50d11a872c45e.

+++ QuantrLogin/app/src/main/java/com/example/quantrlogin/data/dbmodels/CorrelationExperiment.java
@@ -20,7 +20,7 @@ public class CorrelationExperiment extends Experiment implements Serializable {
         this.correlation = correlation;
     }
 
-    public double getCorrelation() {
+    public float getCorrelation() {
         return correlation;
     }
